Update relationships to handle new embedded info
now both the embedded hasMany and embedded belongsTo will pull from the reference cache,
and if it exists they will update the existing record and use it (unless newRecord
is specified in the meta options).  Then regardless of if the record
exists in the cache it will just create a new record and inject it with
the embedded information.

Additionally records which are created (without id), then loaded
afterward (with id/primarykey), will properly inject themselves into the
referenceCache.